完善客户以及标签的相关Service方法

+++ iahms-hotel/src/main/java/cn/apkr/hotel/service/impl/HotelTagServiceImpl.java
@@ -3,11 +3,8 @@ package cn.apkr.hotel.service.impl;
 import java.util.List;
 
 import cn.apkr.common.exception.ServiceException;
-import cn.apkr.common.utils.DateUtils;
 import cn.apkr.common.utils.StringUtils;
-import cn.apkr.hotel.domain.HotelTagGroup;
 import cn.apkr.hotel.mapper.HotelTagGroupMapper;
-import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +13,6 @@ import cn.apkr.hotel.mapper.HotelTagMapper;
 import cn.apkr.hotel.domain.HotelTag;
 import cn.apkr.hotel.service.IHotelTagService;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.validation.annotation.Validated;
 
 /**
  * 标签Service业务层处理
@@ -36,6 +32,17 @@ public class HotelTagServiceImpl implements IHotelTagService {
     private HotelTagGroupMapper hotelTagGroupMapper;
 
     /**
+     * 查询标签列表
+     *
+     * @param hotelTag 标签
+     * @return 标签
+     */
+    @Override
+    public List<HotelTag> selectHotelTagList(HotelTag hotelTag) {
+        return hotelTagMapper.selectHotelTagList(hotelTag);
+    }
+
+    /**
      * 查询标签
      * 
      * @param tagId 标签主键
@@ -47,14 +54,13 @@ public class HotelTagServiceImpl implements IHotelTagService {
     }
 
     /**
-     * 查询标签列表
-     * 
-     * @param hotelTag 标签
-     * @return 标签
+     * 根据标签组ID查询标签列表
+     * @param tagGroupId 标签组ID
+     * @return 标签集合
      */
     @Override
-    public List<HotelTag> selectHotelTagList(HotelTag hotelTag) {
-        return hotelTagMapper.selectHotelTagList(hotelTag);
+    public List<HotelTag> selectHotelTagByTagGroupId(Long tagGroupId) {
+        return hotelTagMapper.selectHotelTagByTagGroupId(tagGroupId);
     }
 
     /**
@@ -68,15 +74,19 @@ public class HotelTagServiceImpl implements IHotelTagService {
     public int insertHotelTag(HotelTag hotelTag) {
         int result = 0;
         // 验证标签名是否冲突
-        if (hotelTagMapper.checkLabelExists(hotelTag.getLabel())) {
-            throw new ServiceException("标签名已存在");
+        String label = hotelTag.getLabel();
+        if (hotelTagMapper.checkLabelExists(label)) {
+            throw new ServiceException(StringUtils.format("标签名 “{}” 已存在", label));
         }
-        if (!hotelTagGroupMapper.checkTagGroupIdExists(hotelTag.getTagGroupId())) {
-            throw new ServiceException("标签组不存在");
-        }
+        // 插入标签
         result = hotelTagMapper.insertHotelTag(hotelTag);
         if (result > 0) {
-            hotelTagGroupMapper.incrementRelTagNum(hotelTag.getTagGroupId());
+            // 插入成功，则更新标签组关联标签数量
+            Long tagGroupId = hotelTag.getTagGroupId();
+            if (tagGroupId != 0 && !hotelTagGroupMapper.checkTagGroupIdExists(tagGroupId)) {
+                throw new ServiceException("标签组不存在");
+            }
+            hotelTagGroupMapper.incrementRelTagNum(tagGroupId);
         }
         return result;
     }
@@ -106,8 +116,8 @@ public class HotelTagServiceImpl implements IHotelTagService {
             // 更新成功，则更新标签组关联标签数量
             Long oldTagGroupId = oldTag.getTagGroupId();
             Long newTagGroupId = hotelTag.getTagGroupId();
-            if (!hotelTagGroupMapper.checkTagGroupIdExists(newTagGroupId)) {
-                throw new ServiceException("标签组不存在");
+            if (newTagGroupId != 0 && !hotelTagGroupMapper.checkTagGroupIdExists(newTagGroupId)) {
+                throw new ServiceException(StringUtils.format("标签组ID {} 不存在", newTagGroupId));
             }
             if (!oldTagGroupId.equals(newTagGroupId)) {
                 hotelTagGroupMapper.decrementRelTagNum(oldTagGroupId);
@@ -115,6 +125,40 @@ public class HotelTagServiceImpl implements IHotelTagService {
             }
         }
         return result;
+    }
+
+    /**
+     * 更新标签关联客户数量
+     * @param tagId 标签ID
+     * @param number 客户数量
+     * @return 结果
+     */
+    @Override
+    public int updateRelCusNum(Long tagId, Integer number) {
+        if (number == null) {
+            number = 0;
+        }
+        return hotelTagMapper.updateRelCusNum(tagId, number);
+    }
+
+    /**
+     * 客户关联数量+1
+     * @param tagId 标签ID
+     * @return 结果
+     */
+    @Override
+    public int incrementRelCusNum(Long tagId) {
+        return hotelTagMapper.incrementRelCusNum(tagId);
+    }
+
+    /**
+     * 客户关联数量-1
+     * @param tagId 标签ID
+     * @return 结果
+     */
+    @Override
+    public int decrementRelCusNum(Long tagId) {
+        return hotelTagMapper.decrementRelCusNum(tagId);
     }
 
     /**
@@ -127,10 +171,11 @@ public class HotelTagServiceImpl implements IHotelTagService {
     @Transactional
     public int deleteHotelTagByTagIds(Long[] tagIds) {
         for (Long tagId : tagIds) {
-            Long tagGroupId = hotelTagMapper.selectHotelTagByTagId(tagId).getTagGroupId();
-            if (hotelTagGroupMapper.checkTagGroupIdExists(tagGroupId)) {
-                hotelTagGroupMapper.decrementRelTagNum(tagGroupId);
+            HotelTag tag = hotelTagMapper.selectHotelTagByTagId(tagId);
+            if (tag == null) {
+                throw new ServiceException(StringUtils.format("标签ID {} 不存在", tagId));
             }
+            hotelTagGroupMapper.decrementRelTagNum(tag.getTagGroupId());
         }
         return hotelTagMapper.deleteHotelTagByTagIds(tagIds);
     }
@@ -144,10 +189,11 @@ public class HotelTagServiceImpl implements IHotelTagService {
     @Override
     @Transactional
     public int deleteHotelTagByTagId(Long tagId) {
-        Long tagGroupId = hotelTagMapper.selectHotelTagByTagId(tagId).getTagGroupId();
-        if (hotelTagGroupMapper.checkTagGroupIdExists(tagGroupId)) {
-            hotelTagGroupMapper.decrementRelTagNum(tagGroupId);
+        HotelTag tag = hotelTagMapper.selectHotelTagByTagId(tagId);
+        if (tag == null) {
+            throw new ServiceException(StringUtils.format("标签ID {} 不存在", tagId));
         }
+        hotelTagGroupMapper.decrementRelTagNum(tag.getTagGroupId());
         return hotelTagMapper.deleteHotelTagByTagId(tagId);
     }
 }
